[BE/refactor] met타입 변경

See merge request s09-bigdata-recom-sub2/S09P22A403!109

+++ backend/fns/src/main/java/ssafy/fns/domain/exercise/service/dto/ExerciseDto.java
@@ -8,6 +8,6 @@ import lombok.Getter;
 public class ExerciseDto {
 
     private Long sportsId;
-    private Long met;
+    private Double met;
     private Long exerciseTime;
 }
